yolo-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
ultralytics==8.1.0
Pillow==10.2.0
opencv-python-headless==4.9.0.80
